requirements.txt
numpy
fastapi
uvicorn
python-multipart
pillow
opencv-python
pyyaml
ultralytics
